Refactor shared types `WalletMeta`, `Account`, `Contact`
