+++ app/components/vault/VaultMasterAccounts.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { smColors } from '../../vars';
 import { DropDown, Tooltip, Dots } from '../../basicComponents';
-import { Account } from '../../types';
+import { Account } from '../../../shared/types';
 
 const DetailsRow = styled.div`
   position: relative;
