+++ app/types/index.ts
@@ -1,4 +1,3 @@
-export { Account, WalletMeta, Contact } from './wallet';
 export { RootState, NetworkState, NodeState, WalletState, UiState, CustomAction, GetState, AppThDispatch } from './redux';
 export { Tx, TxState } from './transactions';
 export { ComputeProvider, ComputeProviders, Reward } from './smesher';
