+++ shared/types/index.ts
@@ -1,0 +1,3 @@
+export * from './node_types.proto';
+export * from './wallet';
+export * from './misc';
