+++ shared/types/node_types.proto.ts
@@ -1,18 +1,3 @@
-export interface NodeVersionAndBuild {
-  version: string;
-  build: string;
-}
-
-export type PublicServices = Array<{
-  url: string;
-  name: string;
-}>;
-
-export type SocketAddress = {
-  ip: string;
-  port: string;
-};
-
 //
 // Rreproduction of some types from `node_types.proto`
 // made by hands
